% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_branch_01.R
\name{brun_branch_01}
\alias{brun_branch_01}
\title{Simulasi Sistem Kejadian Diskret dengan Percabangan (kasus 1)}
\usage{
brun_branch_01(
  L_opr,
  L_mesinA,
  L_mesinB,
  chek = c(0.2, 0.1),
  rsc = c(2, 1, 1),
  rn = 70,
  ongoing = FALSE
)
}
\arguments{
\item{L_opr}{adalah berbentuk fungsi yang menentukan lamanya waktu proses material di operator}

\item{L_mesinA}{adalah berbentuk fungsi yang menentukan lamanya waktu proses material di mesinA}

\item{L_mesinB}{adalah berbentuk fungsi yang menentukan lamanya waktu proses material di mesinB}

\item{chek}{sebuah vektor numerik yang nilai-nilainya antara 0 dan 1, yang memberikan besarnya peluang pada percabangan di lintasan}

\item{rsc}{adalah sebuah vektor numerik untuk memberikan nilai besarnya kapasitas sumber-sumber daya}

\item{rn}{adalah sebuah nilai numerik yang meberikan rentang waktu simulasi, nilai terpasangnya adalah 70}

\item{ongoing}{sebuah nilai logical, yang mana nilai TRUE, berarti mencatat kedatangan entiti yang masih dalam proses di sistem}
}
\value{
\strong{dataset} Arrival dan Resources
}
\description{
Sebuah proses manufaktur dilakukan dalam dua tahap, pertama-tama material
diperiksa oleh seorang operator, jika material yang diperiksa kondisinya baik
maka akan diteruskan untuk diproses dengan mesinA, tetapi bila material kondisinya rusak,
maka akan diteruskan ke mesin B untuk dilakukan koreksi, peluang sebuah material itu
rusak adalah 20%. Setelah selesai dari mesinB, material kembali diperiksa oleh
operator, bila material berhasil dikoreksi, maka akan diteruskan ke mesinA. Peluang
keberhasilan proses koreksi oleh mesinB adalah 90%, material yang tidak berhasil
dikoreksi akan dibuang sebagai limbah \emph{scrap}. Lamanya waktu pemeriksaan oleh
operator menyebar \strong{Triangular(min=0.2, max=0.8, mode=0.4)}. Lamanya waktu
proses oleh mesinA adalah \strong{Uniform(1.5, 2.0)}, sedangkan lamanya waktu
pengkoreksian oleh mesinB adalah \strong{Uniform(2.0, 3.0)}. Pabrik tersebut menyediakan
dua orang operator, satu unit mesinA dan satu unit mesinB. Material tiba untuk
diperoses setiap 2 menit.
}
\examples{
brun_branch_01(function() rtri(1, 0.2, 0.8, 0.4), function() runif(1, 1.5, 2.0) ,
               function() runif(1, 2.0, 3.0))
}
