% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_branch_02.R
\name{brun_branch_02}
\alias{brun_branch_02}
\title{Simulasi Sistem Kejadian Diskret dengan Percabangan (kasus 2)}
\usage{
brun_branch_02(
  L_mesinB,
  ak,
  chek = 0.3,
  rsc = c(1, 1),
  rn = 300,
  ongoing = FALSE
)
}
\arguments{
\item{L_mesinB}{adalah berbentuk fungsi yang menentukan lamanya waktu proses material di mesinB}

\item{ak}{adalah berbentuk fungsi yang memberikan nilai waktu antar kedatangan material}

\item{chek}{adalah nilai numerik lebih besar dari 0 dan kurang dari 1 yang merupakan peluang material masuk ke mesinB, nilai terpasangnya asalah 0.30}

\item{rsc}{adalah vektor numerik yang memberikan kapasitas untuk mesinA dan mesinB, nilai terpasangnya adalah c(1,1)}

\item{rn}{adalah sebuah nilai numerik yang memberikan rentang waktu simulasi, dengan nilai terpasang 300}

\item{ongoing}{sebuah nilai logical, yang mana nilai TRUE, berarti mencatat kedatangan entiti yang masih sedang diproses di sistem.  Hal ini ditunjukkan oleh nilai kolom finished FALSE}
}
\value{
\strong{dataset} Arrivals, Resources, dan Attributes
}
\description{
Suatu sistem pengolahan makanan melakukan pemrosesan dari material mentah sampai
pengemasan (produk siap pasar). Material yang datang bisa langsung masuk antrian
mesinA, apabila antrian di mesinA ≤ 5, dengan lama waktu pemrosesannya diset
sama dengan 6 menit. Jika antrian di mesinA > 5 tetapi < 9 dan antrian di mesinB <= 2,
maka lamanya pemrosesan di mesinA adalah sama dengan 8 menit. Apabila salah satu dari
kedua kondisi di atas tidak terpenuhi, material dikembalikan (tolak) .
Material yang telah selesai diproses oleh mesinA, 70% diantaranya berhasil menjadi
produk yang siap disitribusikan ke pasar. Sedangkan 30% harus masuk ke antrian
mesinB untuk pengolahan lebih lanjut, dengan lama proses menyebar \strong{Exponensial
(rate=0.1 per menit)}. Waktu antar kedatangan material adalah 5 menit, jumlah
mesinA dan mesin B yang tersedia adalah masing-masing 1 unit. Simulasi
dijalankan sampai 300 menit.
}
\examples{
brun_branch_02(function() rexp(1, 0.1), function() 5)
}
