% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_pert_prediksi.R
\name{brun_pert_prediksi}
\alias{brun_pert_prediksi}
\title{Prediksi waktu penyelesaian proyek dengan pendekatan Simulasi Kejadian Diskret.}
\usage{
brun_pert_prediksi(n = 100)
}
\arguments{
\item{n}{suatu numerik, nilai n ini menentukan banyaknya replikasi.}
}
\value{
\strong{dataset} waktu penyelesaian proyek per ulangan

Ringkasan Statistik

Selang_Kepercayaan
}
\description{
Silahkan baca link berikut ini:
\href{https://binusianorg-my.sharepoint.com/personal/i_yudistira_binus_ac_id/_layouts/15/guestaccess.aspx?docid=03e28f10006114071bb8a92e45815a86e&authkey=AV72oobwHx-nMnz1CrfFCog&e=eLyBJ2}{Deskripsi}
}
\examples{
brun_pert_prediksi(n = 100)
}
