% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_sdhn_01.R
\name{brun_sdhn_01}
\alias{brun_sdhn_01}
\title{Sistem Antrian Sederhana.}
\usage{
brun_sdhn_01(ak = 0.4, lm = 0.45, rd = 3, rsc = 1, rn = 50, ongoing = TRUE)
}
\arguments{
\item{ak}{numerik yang merupakan laju kedatangan entiti.}

\item{lm}{numerik yang merupakan laju lamanya layanan.}

\item{rd}{numerik untuk nilai pembulatan.}

\item{rsc}{numerik yang memberikan kapasitas sumberdaya yang tersedia.}

\item{rn}{numerik rentang waktu simulasi.}

\item{ongoing}{= TRUE mencatat kedatangan entiti yang masih dalam proses di sistem.}
}
\value{
data set arrivals
}
\description{
Kampus Binus mempunyai sebuah mesin ATM yang melayani nasabah untuk transaksi perbankan.
Waktu antar kedatangan pengguna ATM adalah merupakan peubah acak yang menyebar
Exponential, dengan rate (lambda : λ = 0.40 kedatangan per jam) dan disimpan dalam objek AK
(dibulatkan 3 digit dibelakang koma). Lamanya waktu yang diperlukan pengguna untuk bertransaksi
pada mesin ATM juga menyebar Eksponesial, dengan rate (lambda : λ = 0.45 pengguna per jam) dan
disimpan dalam objek Lama (dibulatkan 3 digit dibelakang koma).
}
\examples{
brun_sdhn_01(ak=0.40, lm=0.45, rd=3, rsc=1, rn=50, ongoing=TRUE)
}
