% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_sdhn_02.R
\name{brun_sdhn_02}
\alias{brun_sdhn_02}
\title{Sistem Kejadian Diskret dengan dua tipe entiti.}
\usage{
brun_sdhn_02(LamaA, LamaB, wkA, wkB, rsc = 1)
}
\arguments{
\item{LamaA}{berbentuk fungsi yang memberikan nilai lamanya proses entiti A.}

\item{LamaB}{berbentuk fungsi yang memberikan nilai lamanya proses entiti B.}

\item{wkA}{vektor numerik terurut naik yang memberikan nilai waktu kedatangan entiti A.}

\item{wkB}{vektor numerik terurut naik yang memberikan nilai waktu kedatangan entiti B.}

\item{rsc}{sebuah nilai numerik yang memberikan kapasitas sumberdaya yang tersedia.}
}
\value{
\strong{dataset} arrivals dan attributes
}
\description{
Suatu pabrik memiliki sebuah mesin pemroses yang digunakan untuk memproses dua jenis
pekerjaan (“job”), diberi nama jobA dan jobB. JobA datang setiap 8 menit sekali, sedangkan
jobB datang setiap setiap 12 menit sekali. Untuk setiap siklus produksi ada 100 jobA yang harus
diselesaikan, dan ada 50 jobB yang harus diselesaikan. Lama waktu yang dibutuhkan untuk
memproses jobA adalah menyebar Eksponensial dengan rate 1/7 per menit, sedangkan jobB
membutuhkan waktu yang juga menyebar eksponensial tetapi dengan rate 1/12 per menit.
}
\examples{
brun_sdhn_02(LamaA = function() rexp(1, 1/7), LamaB = function() rexp(1, 1/12),
             wkA = cumsum(rep(8, 100)), wkB = cumsum(rep(12, 50)))
}
