% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_sdhn_03.R
\name{brun_sdhn_03}
\alias{brun_sdhn_03}
\title{Sistem Kejadian Diskret dengan lebih dari satu server serial.}
\usage{
brun_sdhn_03(L_opr, L_mesin, AK, src = c(1, 1), rn = 100)
}
\arguments{
\item{L_opr}{berbentuk fungsi yang memberikan nilai lamanya proses di operator.}

\item{L_mesin}{berbentuk fungsi yang memberikan nilai lamanya proses di mesin.}

\item{AK}{berbentuk fungsi nilai waktu antar kedatanagan entiti.}

\item{rn}{numerik yang memberikan rentang waktu simulasi.}

\item{rsc}{numerik yang memberikan kapasitas sumberdaya yang tersedia untuk operator dan mesin.}
}
\value{
dataset Arrivals dan Resources
}
\description{
Sebuah proses manufaktur dilakukan dalam dua tahap, pertama-tama material diperiksa oleh
seorang operator, dengan lama waktu pemeriksaan menyebar Triangular(min=0.2, max=0.8,
mode=0.4). Setelah selesai diperiksa, material diproses oleh sebuah mesin dengan lama
pemrosesan menyebar Uniform(1.5, 2.0). Pabrik tersebut hanya menyediakan satu orang
operator dan satu unit mesin pemroses.
}
\examples{
brun_sdhn_03(L_opr = function() rtri(1, 0.2, 0.8, 0.4), L_mesin = function() runif(1, 1.5, 2.0),
             AK = function() 2, src = c(1, 1), rn=100)
}
