% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_sirkular_01.R
\name{brun_sirkular_01}
\alias{brun_sirkular_01}
\title{Simulasi Melingkar (Sirkular)}
\usage{
brun_sirkular_01(
  MUAT,
  BONGKAR,
  TravelMB,
  TravelBM,
  rsc = c(1, 1),
  rn = 24,
  wkA = at(0),
  wkB = at(0),
  ongoing = FALSE
)
}
\arguments{
\item{MUAT}{sebuah fungsi yang memberikan nilai untuk lamanya aktivitas MUAT di pabrik}

\item{BONGKAR}{sebuah fungsi yang memberikan nilai untuk lamanya aktivitas BONGKAR di gudang}

\item{TravelMB}{sebuah fungsi yang memberikan nilai untuk lamanya waktu perjalanan dari pabrik (MUAT) ke gudang (BONGKAR)}

\item{TravelBM}{sebuah fungsi yang memberikan nilai untuk lamanya waktu perjalanan dari gudang (BONGKAR) ke pabrik (MUAT)}

\item{rsc}{sebuah vektor numerik yang memeberikan besarnya kapasitas tempat MUAT di pabrik dan tempat BONGKAR di gudang}

\item{rn}{sebuah nilai numerik yang memberikan rentang waktu simulasi, nilai terpasangnya adalh 24}

\item{wkA}{adalah sebuah fungsi yang memberikan waktu kedatangan TrukA, nilai terpasangnya adalah at(0)}

\item{wkB}{adalah sebuah fungsi yang memberikan waktu kedatangan TrukB, nilai terpasangnya adalah at(0)}

\item{ongoing}{sebuah nilai logical, yang mana nilai TRUE berarti mencatat kedatangan entiti yang masih sedang diproses di sistem.  Hal ini ditunjukkan oleh nilai kolom finished FALSE}
}
\value{
\strong{dataset} Arrivals
}
\description{
Sebuah perusahaan melakukan pengiriman barang dari pabrik ke gudang.
Perusahaan tersebut menggunakan dua buah truk yang identitik (TrukA dan TrukB) untuk melakukan
aktivitas tersebut. Truk melakukan proses muat di pabrik, yang
membutuhkan waktu 1 jam, kemudian berangkat ke gudang dengan waktu tempuh
1.5 jam. Sampai di gudang, dilakukan proses bongkar, yang membutuhkan waktu
0.5 jam. Selesai proses bongkar di gudang, truk kembali ke pabrik dengan lama
waktu tempuh 1 jam. Simulasi dijalankan sampai waktu 24 jam.
}
\examples{
brun_sirkular_01(MUAT = function() 1, BONGKAR = function() 0.5, TravelMB = function() 1.5,
                           TravelBM = function() 1, rsc=c(1, 1), rn=24, wkA = at(0), wkB = at(0),
                           ongoing=FALSE)

}
