% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_tambang.R
\name{brun_tambang}
\alias{brun_tambang}
\title{Simulasi Melingkar (Sirkular) : Kasus Perusahaan Pertambangan}
\usage{
brun_tambang(
  MUAT20,
  MUAT50,
  BONGKAR20,
  BONGKAR50,
  TravelSC20,
  TravelSC50,
  TravelCS20,
  TravelCS50,
  Beban20,
  Beban50,
  rsc_shovel = c(1, 2),
  rsc_crusher = 1,
  rn = 1200,
  ongoing = FALSE
)
}
\arguments{
\item{MUAT20}{sebuah fungsi yang memberikan lamanya proses muat Truk 20 ton di shovel}

\item{MUAT50}{sebuah fungsi yang memberikan lamanya proses muat Truk 50 ton di shovel}

\item{BONGKAR20}{sebuah fungsi yang memberikan lamanya proses bongkar Truk 20 ton di crusher}

\item{BONGKAR50}{sebuah fungsi yang memberikan lamanya proses bongkar Truk 50 ton di crusher}

\item{TravelSC20}{sebuah fungsi yang memberikan lamanya perjalanan Truk 20 ton menuju crusher}

\item{TravelSC50}{sebuah fungsi yang memberikan lamanya perjalanan Truk 50 ton menuju crusher}

\item{TravelCS20}{sebuah fungsi yang memberikan lamanya perjalanan Truk 20 ton menuju shovel}

\item{TravelCS50}{sebuah fungsi yang memberikan lamanya perjalanan Truk 50 ton menuju shovel}

\item{Beban20}{sebuah fungsi yang memberikan realisasi Beban muat Truk 20 ton}

\item{Beban50}{sebuah fungsi yang memberikan realisasi Beban muat Truk 50 ton}

\item{rsc_shovel}{sebuah vektor numerik yang meberikan kapasitas shovel50 dan shovel20, nilai terpasangnya c(1,2)}

\item{rsc_crusher}{sebuah nilai numerik yang memberikan kapasitas crusher, nilai terpasangnya 1}

\item{rn}{sebuah nilai numerik yang merupakan rentang waktu simulasi}

\item{ongoing}{sebuah nilai logical, yang mana nilai TRUE, berarti mencatat kedatangan entiti yang masih sedang diproses di sistem.  Hal ini ditunjukkan oleh nilai kolom finished FALSE}
}
\value{
\strong{dataset} Arrivals dan Attributes
}
\description{
Sebuah perusahaan tambang mengoperasikan truk, untuk mengakut bijih tambang
dari tiga tempat penambangan ("shovels") ke sebuah "crusher". Ada dua tipe truk
berbeda yang dioperasikan, yaitu truk 20 ton dan truk 50 ton. Ukuran truk ini
mempengaruhi lamanya waktu yang diperlukan untuk loading di "shovel", waktu
tempuh ke "crusher", lamanya waktu bongkar muat di "crusher" dan perjalanan kembali
ke "shovel". Ukuran truk juga mempengaruhi destinasi shovelnya, dari tiga "shovel"
yang tersedia, dua "shovel" diperuntukkan bagi truk 20 ton, dan satu "shovel" untuk truk
50 ton. Disipilin antrian di "shovel" adalah FIFO. Sedangkan antrian di "crusher" berdasarkan
ukuran truk, prioritas diberikan pada truk dengan ukuran besar. Operasi tambang ini
menggunakan 6 truk berukuran 20 ton dan 3 truk berukuran 50 ton. Nilai-nilai input
untuk truk 20 ton adalah sebagai berikut: a) loading time adalah \strong{Exponesial} dengan
\strong{rate 1/5 per menit}, b) waktu tempuh dari "shovel" ke "crusher" adalah 150 menit,
c) lamanya proses bongkar di "crusher" adalah \strong{Exponesial(1/2)}, d) waktu tempuh
kembali ke "shovel" adalah 90 menit. Sedangkan nilai-nilai input untuk truk 50 ton
adalah sebagai berikut: a) loading time adalah \strong{Exponesial(1/10)}, b) waktu tempuh
dari "shovel" ke "crusher" adalah 180 menit, c) lamanya proses bongkar di "crusher"
adalah \strong{Exponesial(1/4)}, d) waktu tempuh kembali ke "shovel" adalah 120 menit.
hitung total bijih yang diangkut dalam 20 jam (1200 menit) kerja tambang. Dimana kapasitas muat
dan realisasi muat, untuk Truk 20 ton adalah 2 ton dan 2 + \strong{Normal(0, 0.05)}.  Sedangkan
untuk Truk 50 ton kapasitasnya adalah 5 ton, sedangkan realisasinya adalah 5 + \strong{Normal(0, 0.1)}
}
\examples{
brun_tambang(MUAT20 = function() rexp(1, 1/5), MUAT50 = function() rexp(1, 1/10),
             BONGKAR20 = function() rexp(1, 1/2), BONGKAR50 = function() rexp(1, 1/4),
             TravelSC20 = function() 150, TravelSC50 = function() 180, TravelCS20 = function() 90,
             TravelCS50 = function() 120, Beban20 = function() round(2 + rnorm(1, 0, 0.05), 2),
             Beban50 = function() round(5 + rnorm(1, 0, 0.1), 2), rsc_shovel=c(1, 2),
             rsc_crusher = 1, rn=1200, ongoing=FALSE)
}
