% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_viz_n_single.R
\name{brun_viz_n_single}
\alias{brun_viz_n_single}
\title{Visualisasi Simulasi Sistem Kejadian Diskret dengan satu Server dan n kali replikasi}
\usage{
brun_viz_n_single(
  AK,
  Lama,
  RUN = 100,
  n = 30,
  graph = c("None", "Utility", "Wait_Time", "Flow_Time", "Total_Serv", "Avg_Flow",
    "Avg_Waiting", "Queue_Rate", "Server_Downtime", "Avg_Queue_Length"),
  seed = as.numeric(Sys.time())
)
}
\arguments{
\item{AK}{adalah sebuah fungsi sebaran peluang / nilai konstan untuk waktu antar kedatangan entiti}

\item{Lama}{adalah sebuah fungsi sebaran peluang /nilai konstan untuk lamanya proses di server}

\item{RUN}{adalah sebuah nilai numerik, yang memberikan rentang waktu simulasi nilai terpasangnya adalah 100}

\item{n}{adalah sebuah nilai numerik, yang memberikan banyaknya replikasi, nilai terpasangnya adalah 30}

\item{graph}{pilih salah satu dari "None", "Queue", "Utility", "Wait_Time", "Flow_Time", "Total_Serv", "Avg_Flow", "Avg_Waiting", "Queue_Rate",
"Server_Downtime","Avg_Queue_Length"}

\item{seed}{sebuah nilai numerik untuk bibit angka acak, nilai terpasangnya adalah as.numeric(sys.time)}
}
\value{
\strong{dataset} arrival, resources,

statistics yang meliputi
\emph{total_serv, avg_flow_time, avg_wait_time, longest_flow_time, longest_waiting_time
last_time, server_downtime, queue_rate, utility_rate, avg_queue_length, avg_entity_in_system}
}
\description{
Penjelasan lebih jauh ada pada link berikut ini
\href{https://binusianorg-my.sharepoint.com/personal/i_yudistira_binus_ac_id/_layouts/15/guestaccess.aspx?docid=02922ecc2fe594582a1bcecdbfbf1d3a1&authkey=AV8ib0HpB_D9aA2RhpkfID4&e=vQxLPz}{Visualisasi}
}
\examples{
brun_viz_n_single(AK = function() round(rexp(n=1, rate=0.40),3),
         Lama = function() round(rexp(n=1, rate=0.45),3), graph="Utility")
}
