% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brun_viz_singleDES.R
\name{brun_viz_single}
\alias{brun_viz_single}
\title{Visualisasi DES output - tanpa replikasi}
\usage{
brun_viz_single(
  AK,
  Lama,
  RUN = 100,
  graph = c("None", "Queue", "Utility", "System", "All"),
  seed = as.numeric(Sys.time())
)
}
\arguments{
\item{AK}{adalah sebuah fungsi yang memberikan waktu antar kedatangan suatu entiti}

\item{Lama}{adalah sebuah fungsi yang memberikan lamanya proses layanan di server}

\item{RUN}{adalah sebuah nilai numerik yang memberikan rentang waktu simulasi, nilai defaultnya adalah 100}

\item{graph}{choose salah satu dari "None", Queue", "Utility", "System", atau "All", untuk menayangkan atau tidak menayangkan grafik}
}
\value{
\strong{dataset} "out_going_output", "resource_ouput", "wide_stat_value","long_stat_value"

\strong{grafik} "Que Length", "Total_Entities_in System" and "Server_Utility"
}
\description{
lihat link:
\href{https://binusianorg-my.sharepoint.com/personal/i_yudistira_binus_ac_id/_layouts/15/guestaccess.aspx?docid=02922ecc2fe594582a1bcecdbfbf1d3a1&authkey=AV8ib0HpB_D9aA2RhpkfID4&e=7tNyZe}{Deskripsi}
}
\examples{
brun_viz_single(AK = function() round(rexp(n = 1, rate = 0.40),3),
          Lama = function() round(rexp(n = 1, rate = 0.45),3), graph = "System")


}
